requirements.txt
flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
dnspython==2.4.2
email-validator==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
